Versioning 2.0.18, Fix for WordPress 3.5 manging wp_enqueue URLs, fixes to edge-cases with widget dependencies, fix to link-builder when entering lat/lng values.
git-svn-id: https://plugins.svn.wordpress.org/dsidxpress/trunk@637401 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ js/admin-utilities.js
@@ -711,14 +711,14 @@ dsIdxLinkBuilder = {
 				{
 				    label: 'Minimum Longitude',
 				    id: 'idx-q-LongitudeMin',
-				    type: 'int',
+				    type: 'decimal',
 				    mode: 'single',
 				    hint: ''
 				},
 				{
 				    label: 'Maximum Longitude',
 				    id: 'idx-q-LongitudeMax',
-				    type: 'int',
+				    type: 'decimal',
 				    mode: 'single',
 				    hint: ''
 				}
